+++ src/main/java/org/ccclll777/alldocsbackend/service/TagService.java
@@ -18,6 +18,12 @@ public class TagService {
     private TagDao tagDao;
     @Autowired
     private FileTagDao fileTagDao;
+
+    /**
+     * 插入标签
+     * @param tag
+     * @return
+     */
     public int insertTag(Tag tag) {
         int count = tagDao.haveTag(tag.getName());
         if(count > 0) {
@@ -25,9 +31,21 @@ public class TagService {
         }
         return tagDao.insertTag(tag);
     }
+
+    /**
+     * 更新标签信息
+     * @param tag
+     * @return
+     */
     public int updateTag(Tag tag){
         return tagDao.updateTag(tag);
     }
+
+    /**
+     * 删除标签
+     * @param tagId
+     * @return
+     */
     public int deleteTag(int tagId) {
         int count = fileTagDao.fileCountByTagId(tagId);
         if (count > 0) {
@@ -45,6 +63,11 @@ public class TagService {
         return tagDao.tagCount();
     }
 
+    /**
+     * 批量删除标签
+     * @param tagIds
+     * @return
+     */
     public int deleteTagList(String tagIds) {
         String[] idxs = tagIds.split(",");
         if(idxs.length == 0){
@@ -52,7 +75,10 @@ public class TagService {
         }
         for (String idx : idxs) {
             Integer id = Integer.parseInt(idx);
-            tagDao.deleteTag(id);
+            int count = fileTagDao.fileCountByTagId(id);
+            if (count <= 0) {
+                tagDao.deleteTag(id);
+            }
         }
         return 1;
     }
